.kiro/rag_backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
sentence-transformers>=2.7.0
faiss-cpu==1.13.2
PyPDF2==3.0.1
numpy>=1.24.3
python-multipart==0.0.6
requests>=2.31.0
